+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "wdio run ./src/core/config/wdio.conf.js",
     "wdio": "wdio run ./wdio.conf.js",
     "test:smoke": "wdio run wdio.conf.js --cucumberOpts.tagExpression='@smoke'"
   },
