+++ src/business/pageobjects/favorites.page.js
