+++ src/business/step-definitions/registration.steps.js
@@ -1,8 +1,8 @@
 const { Given, When, Then } = require("@wdio/cucumber-framework");
-const HomePage = require("../pageobjects/home.page");
-const LoginPage = require("../pageobjects/login.page");
+const HomePage = require("../pageobjects/home.page.js");
+const LoginPage = require("../pageobjects/login.page.js");
 const RegisterPage = require("../pageobjects/register.page.js");
-const AccountPage = require("../pageobjects/account.page");
+const AccountPage = require("../pageobjects/account.page.js");
 const { validUser } = require("../data/userData.js");
 
 Given("the user opens the website", async () => {
